requirements.txt
fastapi
uvicorn
python-dotenv
requests
streamlit
mariadb
ollama
chromadb
sentence-transformers